chore: update tyings
