+++ src/frontend/screens/ObservationFields/Number.tsx
@@ -1,29 +1,26 @@
 import * as React from 'react';
 import {StyleSheet, TextInput} from 'react-native';
-import {QuestionLabel} from './QuestionLabel';
-import {Field} from '@comapeo/schema';
-import {
-  useDraftObservationActions,
-  useDraftObservationState,
-} from '../../contexts/DraftObservationContext';
+import {Observation} from '@comapeo/schema';
 
-export const Number = React.memo<{field: Field}>(({field}) => {
-  const tags = useDraftObservationState(store => store.value?.tags);
-  const {updateTag} = useDraftObservationActions();
-  const value = tags ? tags[field.tagKey] : '';
+type NumberProps = {
+  updateTag: (value: number) => void;
+  tagValue?: Observation['tags'][number];
+};
+
+export const Number = ({updateTag, tagValue}: NumberProps) => {
   return (
-    <React.Fragment>
-      <QuestionLabel field={field} />
+    <>
       <TextInput
         testID="OBS.details-inp"
-        value={typeof value === 'string' ? value : ''}
+        value={typeof tagValue === 'number' ? String(tagValue) : ''}
         onChangeText={newVal =>
           updateTag(
-            field.tagKey,
-            newVal
-              .replace(/[^0-9.-]/g, '') // Allow digits, decimal, and negative sign
-              .replace(/(?!^)-/g, '') // Remove any minus sign that is not at the start
-              .replace(/(\..*?)\./g, '$1'), // Remove additional decimal points
+            parseFloat(
+              newVal
+                .replace(/[^0-9.-]/g, '') // Allow digits, decimal, and negative sign
+                .replace(/(?!^)-/g, '') // Remove any minus sign that is not at the start
+                .replace(/(\..*?)\./g, '$1'), // Remove additional decimal points
+            ),
           )
         }
         keyboardType="numeric"
@@ -34,9 +31,9 @@ export const Number = React.memo<{field: Field}>(({field}) => {
         textContentType="none"
         autoFocus
       />
-    </React.Fragment>
+    </>
   );
-});
+};
 
 const styles = StyleSheet.create({
   textInput: {
