+++ src/frontend/screens/ObservationFields/SelectOne.tsx
@@ -1,64 +1,41 @@
 import * as React from 'react';
-import {View, StyleSheet} from 'react-native';
-import {Text} from '../../sharedComponents/Text';
+import {View, StyleSheet, TouchableOpacity} from 'react-native';
 import MaterialIcon from '@react-native-vector-icons/material-icons';
+import {HeaderText} from '../../sharedComponents/Text/HeaderText';
+import {Field, Observation} from '@comapeo/schema';
 
-import {TouchableNativeFeedback} from '../../sharedComponents/Touchables';
-import {VERY_LIGHT_BLUE} from '../../lib/styles';
-import {QuestionLabel} from './QuestionLabel';
-
-import type {QuestionProps} from './Question';
-import {ViewStyleProp} from '../../sharedTypes';
-import {SelectOneField} from '../../sharedTypes/PresetTypes';
-import {
-  useDraftObservationActions,
-  useDraftObservationState,
-} from '../../contexts/DraftObservationContext';
-
-interface Props extends QuestionProps {
-  field: SelectOneField;
-}
-
-type RadioItemProps = {
-  checked: boolean;
-  onPress: () => void;
-  label: string;
-  style: ViewStyleProp;
+type SelectOneProps = {
+  options: NonNullable<Field['options']>;
+  updateTag: (value: Observation['tags'][number]) => void;
+  tagValue?: Observation['tags'][number];
 };
 
-const RadioItem = ({checked, onPress, label, style}: RadioItemProps) => (
-  <TouchableNativeFeedback
-    onPress={onPress}
-    background={TouchableNativeFeedback.Ripple(VERY_LIGHT_BLUE, false)}>
-    <View style={style}>
-      <MaterialIcon
-        name={checked ? 'radio-button-checked' : 'radio-button-unchecked'}
-        size={30}
-      />
-      <Text style={styles.itemLabel}>{label}</Text>
-    </View>
-  </TouchableNativeFeedback>
-);
-
-export const SelectOne = React.memo<Props>(({field}) => {
-  const {updateTag} = useDraftObservationActions();
-  const tags = useDraftObservationState(state => state.value?.tags);
-
+export const SelectOne = ({options, updateTag, tagValue}: SelectOneProps) => {
   return (
     <>
-      <QuestionLabel field={field} />
-      {field.options.map((item, index) => (
-        <RadioItem
+      {options.map((item, index) => (
+        <TouchableOpacity
           key={item.label}
-          onPress={() => updateTag(field.tagKey, item.value)}
-          checked={tags && item.value === tags[field.tagKey] ? true : false}
-          label={item.label}
-          style={[styles.radioContainer, index === 0 ? styles.noBorder : {}]}
-        />
+          onPress={() => updateTag(item.value)}>
+          <View
+            style={[styles.radioContainer, index === 0 ? styles.noBorder : {}]}>
+            <MaterialIcon
+              name={
+                item.value === tagValue
+                  ? 'radio-button-checked'
+                  : 'radio-button-unchecked'
+              }
+              size={30}
+            />
+            <HeaderText style={styles.itemLabel} variant="header4">
+              {item.label}
+            </HeaderText>
+          </View>
+        </TouchableOpacity>
       ))}
     </>
   );
-});
+};
 
 const styles = StyleSheet.create({
   radioContainer: {
@@ -73,11 +50,7 @@ const styles = StyleSheet.create({
     borderTopWidth: 0,
   },
   itemLabel: {
-    fontSize: 18,
-    lineHeight: 24,
     marginLeft: 20,
     flex: 1,
-    color: 'black',
-    fontWeight: '700',
   },
 });
